+++ src/Geocoder/Provider/IpInfoDb.php
@@ -14,7 +14,7 @@ use Geocoder\Exception\InvalidArgument;
 use Geocoder\Exception\InvalidCredentials;
 use Geocoder\Exception\NoResult;
 use Geocoder\Exception\UnsupportedOperation;
-use Geocoder\GeocoderResult;
+use Geocoder\Collection;
 use Http\Client\HttpClient;
 
 /**
@@ -118,7 +118,7 @@ final class IpInfoDb extends AbstractHttpProvider implements Provider
     /**
      * @param string $query
      *
-     * @return GeocoderResult
+     * @return Collection
      */
     private function executeQuery($query)
     {